+++ agent/src/agent.ts
@@ -2,6 +2,7 @@ import path from 'path'
 
 import * as vscode from 'vscode'
 
+import { convertGitCloneURLToCodebaseName } from '@sourcegraph/cody-shared/dist/utils'
 import { Client, createClient } from '@sourcegraph/cody-shared/src/chat/client'
 import { registeredRecipes } from '@sourcegraph/cody-shared/src/chat/recipes/agent-recipes'
 import { SourcegraphNodeCompletionsClient } from '@sourcegraph/cody-shared/src/sourcegraph-api/completions/nodeClient'
@@ -13,13 +14,13 @@ import { AgentTextDocument } from './AgentTextDocument'
 import { newTextEditor } from './AgentTextEditor'
 import { AgentWorkspaceDocuments } from './AgentWorkspaceDocuments'
 import { AgentEditor } from './editor'
-import { MessageHandler } from './jsonrpc'
-import { AutocompleteItem, ClientInfo, ExtensionConfiguration, RecipeInfo } from './protocol'
+import { MessageHandler } from './jsonrpc-alias'
+import { AutocompleteItem, ClientInfo, ExtensionConfiguration, RecipeInfo } from './protocol-alias'
 import * as vscode_shim from './vscode-shim'
 
 const secretStorage = new Map<string, string>()
 
-function initializeVscodeExtension(): void {
+export function initializeVscodeExtension(): void {
     activate({
         asAbsolutePath(relativePath) {
             return path.resolve(process.cwd(), relativePath)
@@ -123,14 +124,14 @@ export class Agent extends MessageHandler {
             this.workspace.setActiveTextEditor(newTextEditor(this.workspace.agentTextDocument(document)))
         })
         this.registerNotification('textDocument/didOpen', document => {
-            this.workspace.setDocument(document)
+            this.workspace.addDocument(document)
             const textDocument = this.workspace.agentTextDocument(document)
             vscode_shim.onDidOpenTextDocument.fire(textDocument)
             this.workspace.setActiveTextEditor(newTextEditor(textDocument))
         })
         this.registerNotification('textDocument/didChange', document => {
             const textDocument = this.workspace.agentTextDocument(document)
-            this.workspace.setDocument(document)
+            this.workspace.addDocument(document)
             this.workspace.setActiveTextEditor(newTextEditor(textDocument))
             vscode_shim.onDidChangeTextDocument.fire({
                 document: textDocument,
@@ -265,6 +266,11 @@ export class Agent extends MessageHandler {
                 console.error('getRepoId', result)
             }
             return typeof result === 'string' ? result : null
+        })
+
+        this.registerRequest('git/codebaseName', ({ url }) => {
+            const result = convertGitCloneURLToCodebaseName(url)
+            return Promise.resolve(typeof result === 'string' ? result : null)
         })
 
         this.registerNotification('autocomplete/clearLastCandidate', async () => {