+++ agent/src/editor.ts
@@ -11,7 +11,7 @@ import {
 
 import { Agent } from './agent'
 import { DocumentOffsets } from './offsets'
-import { TextDocument } from './protocol'
+import { TextDocument } from './protocol-alias'
 
 export class AgentEditor implements Editor {
     public controllers?: ActiveTextEditorViewControllers | undefined
@@ -52,7 +52,7 @@ export class AgentEditor implements Editor {
 
     public getActiveTextEditorSelection(): ActiveTextEditorSelection | null {
         const document = this.activeDocument()
-        if (document === undefined || document.content === undefined || document.selection === undefined) {
+        if (document?.content === undefined || document.selection === undefined) {
             return null
         }
         const offsets = new DocumentOffsets(document)