+++ vscode/src/jsonrpc/agent-protocol.ts
@@ -4,7 +4,7 @@ import { RecipeID } from '@sourcegraph/cody-shared/src/chat/recipes/recipe'
 import { ChatMessage } from '@sourcegraph/cody-shared/src/chat/transcript/messages'
 import { event } from '@sourcegraph/cody-shared/src/sourcegraph-api/graphql/client'
 
-import { CompletionEvent } from '../../vscode/src/completions/logger'
+import { CompletionEvent } from '../completions/logger'
 
 // This file documents the Cody Agent JSON-RPC protocol. Consult the JSON-RPC
 // specification to learn about how JSON-RPC works https://www.jsonrpc.org/specification
@@ -43,6 +43,8 @@ export type Requests = {
 
     'graphql/getRepoIdIfEmbeddingExists': [{ repoName: string }, string | null]
     'graphql/getRepoId': [{ repoName: string }, string | null]
+
+    'git/codebaseName': [{ url: string }, string | null]
 
     // ================
     // Server -> Client