+++ vscode/src/main.ts
@@ -113,13 +113,13 @@ const register = async (
     const authProvider = new AuthProvider(initialConfig)
     await authProvider.init()
 
-    const symfRunner = platform.createSymfRunner?.(context, initialConfig.accessToken)
+    const symfRunner = platform.createSymfRunner?.(context, initialConfig.serverEndpoint, initialConfig.accessToken)
     if (symfRunner) {
         authProvider.addChangeListener(async (authStatus: AuthStatus) => {
             if (authStatus.isLoggedIn) {
-                symfRunner.setAuthToken(await getAccessToken())
+                symfRunner.setSourcegraphAuth(authStatus.endpoint, await getAccessToken())
             } else {
-                symfRunner.setAuthToken(null)
+                symfRunner.setSourcegraphAuth(null, null)
             }
         })
     }
@@ -465,14 +465,18 @@ const register = async (
             completionsProvider.dispose()
         }
 
-        completionsProvider = await createInlineCompletionItemProvider({
-            config,
-            client: codeCompletionsClient,
-            statusBar,
-            contextProvider,
-            authProvider,
-            triggerNotice: notice => sidebarChatProvider.triggerNotice(notice),
-        })
+        completionsProvider = await createInlineCompletionItemProvider(
+            {
+                config,
+                client: codeCompletionsClient,
+                statusBar,
+                contextProvider,
+                authProvider,
+                triggerNotice: notice => sidebarChatProvider.triggerNotice(notice),
+            },
+            context,
+            platform
+        )
     }
     // Reload autocomplete if either the configuration changes or the auth status is updated
     vscode.workspace.onDidChangeConfiguration(event => {
@@ -519,7 +523,7 @@ const register = async (
             externalServicesOnDidConfigurationChange(newConfig)
             void createOrUpdateEventLogger(newConfig, isExtensionModeDevOrTest)
             platform.onConfigurationChange?.(newConfig)
-            symfRunner?.setAuthToken(newConfig.accessToken)
+            symfRunner?.setSourcegraphAuth(newConfig.serverEndpoint, newConfig.accessToken)
         },
     }
 }