+++ vscode/test/completions/mock-vscode.ts
@@ -26,13 +26,16 @@ const vscodeMock = {
                             return true
                         case 'cody.serverEndpoint':
                             return 'https://sourcegraph.com/'
-                        // case 'cody.autocomplete.advanced.provider':
-                        //     return 'unstable-fireworks'
+                        case 'cody.autocomplete.advanced.provider':
+                            return 'anthropic'
                         // case 'cody.autocomplete.advanced.model':
                         //     return 'llama-code-13b'
                         default:
                             return undefined
                     }
+                },
+                update() {
+                    return undefined
                 },
             }
         },